Se não forem recebidos pelo menos 40% dos pacotes de um ficheiro, será pedido o ficheiro todo de novo de forma a reduzir o flooding de pedidos e a consequente criação de threads no fileSender

+++ P2P/src/main/java/files/Ficheiro.java
@@ -79,6 +79,8 @@ public class Ficheiro {
         int i;
         String folderPath = "NODE_" + this.nodeID + "/" + folder + "/" + this.fileName;
         File ficheiro = new File(folderPath);
+        File filePointer;
+        String path;
 
         while((!ficheiro.exists() && !ficheiro.isDirectory()) && !ficheiro.mkdir());
 
@@ -86,8 +88,14 @@ public class Ficheiro {
         for(i = 0; i < size; i++){
             try {
                 FileChunk fc = fcs[i];
-                file = Paths.get(folderPath + "/" + fc.getPlace() + ".filechunk");
-                Files.write(file, fc.getFileChunk());
+                path = folderPath + "/" + fc.getPlace() + ".filechunk";
+                filePointer = new File(path);
+
+                //SO VAI ESCREVER O FICHEIRO SE ELE NAO EXISTIR
+                if(!filePointer.exists()) {
+                    file = Paths.get(path);
+                    Files.write(file, fc.getFileChunk());
+                }
             }
                 catch (Exception e) {
                 e.printStackTrace();
